Lower memory usage when parsing heap dumps on M
Parsing an hprof file generated on M results in a Snapshot filled with duplicate GC root entries.  This extra memory usage leaves little room for LeakCanary to analyze and find the leak trace and often results in an OutOfMemoryError.  This commit removes the duplicate GC root entries to alleviate the memory pressure.

Fixes #223