Ignore computeIgnoredBitmapRetainedSize on O preview
In O preview heap dumps, android.graphics.Bitmap no longer has a field
named mBuffer.  In certain cases when a Bitmap was part of a leak trace,
mBuffer was used to update the leaking instance's total retained size.

This commit is a workaround until sources for Android O are published.

Fixes #759