Moving LeakDirectoryProvider to LeakCanaryInternals
On top of the AndroidHeapDumper, we also need to access LeakDirectoryProvider from DisplayLeakActivity. We'll soon need to also access it from the UI test run listener. We can't keep the instance as a field of the RefWatcher singleton because DisplayLeakActivity lives in a different process.

This change renames LeakCanary.setDisplayLeakActivityDirectoryProvider to LeakCanary.setLeakDirectoryProvider and makes sure we use that global LeakDirectoryProvider everywhere.