Import hprof for analysis (#1305)
Also removed the wait for pending files to be removed. Since we have a threshold now, we won't be dumping the heap as often, and we're still using a serial service for the analysis so things will enqueue nicely. This change allows us to reuse the same mechanism for importing hprof even if an analysis is in progress.

Fixes #1040