Pending analysis won't block heap dumps anymore (#1315)
Previously, we looked for a `*_pending.hprof` file younger than 10 minutes on the filesystem to detect if an analysis was in progress, and skip heap dumps to avoid constantly doing heap dumps.

This meant that if the process died (e.g. crash) during an analysis, no heap dumps would happen in the next 10 minutes.

Heap analysis is now fast, it finds many leaks at once, it's serial and in a single process. This means we can remove this behavior and just dump the heap whenever thresholds are met and then enqueue work to the serial service.

If the process is killed during an analysis, it won't resume but it's easy enough to import the hprof again, from the sd card.

Fixes #1312