Adding back support for legacy LeakCanary hprof files
This comes for cheap and means we can analyze any kind of hprof. Moving forward we'll keep compatibility of LeakCanary hprof parsing.

Also moved the Android configuration into the analyzer repo and broke their dependency from the Android classes. Goal is to run a properly configured analyzer from the command line.

Note: this commit comes as no extra cost in size as the hprof files already existed in the git repo.