Stop renaming heap dumps
Rather than having a default behavior of deleting the pending heap dump with a default listener that renames it so that it's not deleted unless you don't use the default.. we just keep it around, with the right name from the get go. Simplifies logic quite a bit and allows for listener delegation and reporting of the heap dump location.

Fixes #1416