Don't crash on missing hprof file during analysis
If the hprof file is missing, return a HeapAnalysisFailure result instead of crashing.

With #1670 we saw this could happen when a lot of leaks are detected and more than 7 heap dumps are enqueued to be processed. 3cf9c8d9dbec80c4f500e91a68b4dfdbf854299b introduced a max of 1 heap dumps per minute, so this edge case should happen less often now.

Fixes #1670