Wait for heap to be done dumping
We've seen several cases where analyzing the heap right after the heap dump fails (missing objects) but retrying soon after succeeds, which would indicate that the heap dump file isn't done being written to disk. This is an attempt at fixing that by sleeping while the file is appended to.