Detect Android X View Model leaks
AndroidXFragmentDestroyWatcher adds a spy view model in every ViewModelStoreOwner instance (ie FragmentActivity and Fragment). ViewModelClearedWatcher
is a view model that holds on to the map of view models backing its store. When ViewModelClearedWatcher receives
the onCleared() callback, it adds each live view model from the store to the ObjectWatcher.

Other changes:

* Renamed FragmentLeakTest to LifecycleLeaksTest.
* Added tests for activity leaks and view model leaks.
* Refactored leak tests for readability.

Fixes #1731