Prioritize system classes
In Android VMs, AFAIK there is only ever one instance of each class. However, classes can become unloaded and loaded again later, if there are no side effects. That's totally transparent to the runtime, however one of the side effects is that the class will be present twice in the heap dump.

Also, on Android, all loaded classes are held by a sticky class GC Root (i.e. they're system classes). So that's how we find which of the classes are loaded vs not when we get duplicated classes.

This change makes it so we always prioritize system classes when finding classes by name, ensure we find the loaded class.

It's worth noting that standard JVMs behave differently, so we need a different behavior for those. The ideal fix here would be to run a mark and sweep and ignore unreachable classes, however that's time consuming (and harder to do prior to having parsed the heap dump).

Fixes #1759
Fixes #2072