Fix crash when going back while already going back (#1780)
The navigation utilities we have in place normally take care of canceling scheduled work (IO and back on main thread as well) when we leave a screen. Unfortunately we were previously waiting for the view animation to be done (view detached). Instead we now cancel the work as soon as we decide to move.

Fixes #1768