Fix for #1081
Uses the fix from @emartynov ([here](https://github.com/square/leakcanary/issues/1081#issuecomment-489694028)) except we use the application context instead of activity context to avoid reintroducing a leak.

This effectively flushes the pool on activity destroy.

Ideally, we'd do that on fragment as well, and pretty much any time a chunk of the view hierarchy changes.