Refactor disabled reason
This PR builds upon #1894 to make the disabling of heap dumping more central. LeakCanary itself stops setting the dumpHeap parameter (which is now only used by developers) and is able to surface why it's not dumping the heap.

Also tightened up strings to limit the volume of logs and clarify the logs.

Fixes #1919