Keep class fields in memory
This reduces IO reads by more than 20% by keeping the fields part of the class dump records in memory. To limit the impact
on retained memory, the fields are stored as the exact bytes that are found in the heap dump. In the example heap dumps
this adds 200 to 300 KB which is reasonable.

Also changed StreamingHprofReader to provide a lower level API that gives a reader for each heap dump tag found. This allow
consumers to read exactly the bytes they need and avoid creating a lot of temporary objects. Previously this was done
by introducing special "skip" record types, but the needs kept changing.

The previous API is still available for convenience: `StreamingHprofReader.readerFor(heapDumpFile).asStreamingRecordReader()`