Fix disk read violation for SharedPrefs
Unfortunately changing the thread policy didn't work as the next immediately call to read a value from the shared pref creates a violation again.

This fixes it by changing the thread from which we trigger the early read, which is probably better anyway because we're not blocking the main thread.

Fixes #1984