Surface toast root view leaks
Toasts that hold on to an activity context should be canceled when the activity is destroyed. If not, they'll leak the activity context through various ways, one of which is via attached root views if they're being displayed.

This change removes an incorrect AndroidReferenceMatchers tag and adds details to ViewRootImpl to help understand this.

Fixes #1982