Add more logging in InstrumentationLeakDetector
I am trying to enable leak canary in tests in my repo's
CI pipeline but found that leak detector for all tests
are returned with NoAnalysis.

So I am adding logs in this commit to help us understand
where the exit point is, given that there are multiple
return statements and checking on the timestamp is not
always reliable.