More natural list selector
List item background is replaced by a ripple on supported APIs.

ListView's listSelector is explicitly overridden with a custom Drawable that restores the focused state which is kind of clunky in ListView.
Unfortunately, using itemsCanFocus with a focusable View inside the list item will prevent the AdapterView.OnItemClickListener to work as intended.