Switch AndroidLeakFixes to HandlerThread
On Android 8 and 9, I found examples in CI where FLUSH_HANDLER_THREADS was not flushing GoogleApiHandler. I couldn't figure out why, but a heap dump revealed that the FLUSH_HANDLER_THREADS wasn't scheduled on the executor's queue. The logs showed that it had been running though. So it looks like scheduling repeated task might be canceled under some conditions.

This change switches us to using a HandlerThread which we know works well.