RandomAccess uses RandomAccessFile
This changes the random access read impl from FileChannel.transferTo to RandomAccessFile.read. Experiments in #2333 showed that this new impl is 40% faster on API 23 due to lock contention, and otherwise 20% faster on API 24. FileChannelImpl.transferTo will use a java.nio.MemoryBlock$MemoryMappedBlock under the hood, which uses Libcore.os.mmap and then has a finalizer method that calls Libcore.os.munmap. This seemed to trigger a lot of contention locking between the Finalizer and the RefQueue daemons.