Infinite ObjectAnimators caused Drawable leaks
Drawable has an `mCallback` field that is a weak ref to the callback view. It's a weak ref as an attempt to prevent leaks ([change](https://cs.android.com/android/_/android/platform/frameworks/base/+/f2a47782f31b58d2d31bd00b50fe43604af8b9c2)). Unfortunately, when an object animator runs a drawable forever, the `mCallback` weak ref will be checked every frame and the callback reference held as a java local for a brief moment, which makes it really hard for the GC to collect the targeted object.

Fixes #2116